cloudrun/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.1
google-cloud-monitoring==2.18.0